requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
et-xmlfile==1.1.0
google==3.0.0
h11==0.13.0
html5lib==1.1
idna==3.3
lxml==4.9.1
openpyxl==3.0.10
outcome==1.2.0
PySocks==1.7.1
PyYAML==6.0
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdrivermanager==0.10.0
webencodings==0.5.1
wsproto==1.2.0
